More structuring Done
