+++ app/forms.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField, IntegerField, BooleanField
 from wtforms.validators import DataRequired, Email, Length, EqualTo, ValidationError
-from models import User
+from app.models import User
 
 class UserRegistration(FlaskForm):
     username = StringField('Username',
