+++ app/models.py
@@ -6,9 +6,9 @@ from app import db, login_manager
 db = SQLAlchemy()
 
 
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(int(user_id))
+# @login_manager.user_loader
+# def load_user(user_id):
+#     return User.query.get(int(user_id))
 
 class User(db.Model, UserMixin):
     __tablename__ = 'user'
